Refactor: Extract superglueReducer tests from reducers.spec.js
Closes #205
